+++ sqlite3/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("cargo:rustc-link-search=native=target/debug");
+}
